app/microservices/auth/requirements.txt
# Core
fastapi>=0.122.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
pydantic-settings>=2.12.0

# Database
sqlmodel>=0.0.27
sqlalchemy>=2.0.44
asyncpg>=0.31.0

# Security
python-jose[cryptography]>=3.5.0
argon2-cffi>=25.1.0

# HTTP Client
httpx>=0.28.0

# Email
requests>=2.32.0

# Utilities
python-multipart>=0.0.20
python-dotenv>=1.2.0
email-validator>=2.3.0
